test(cli): split runner help coverage
